+++ core/src/types.rs
@@ -8,8 +8,8 @@ pub struct PrivateGraphChunk {
 	/// User-Assigned Key Identifier
 	pub prids: Vec<DsnpPrid>,
 
-	/// Unix epoch seconds when PRIds for this chunk were last refreshed, rounded to the nearest 1000
-	pub last_updated: u64,
+	/// Days since the Unix Epoch when PRIds for this chunk were last refreshed
+	pub last_updated: u32,
 
 	/// connections
 	pub inner_graph: DsnpInnerGraph,
